Rename sideMenu.ts to SideMenu.ts
To follow coding style: use pascal case if the file contains a single
class which is the default export.
